refactor: Winner 객체 추가
